refactor: move ScalingUpApplier to state.appliers and fix golden file
`ScalingUpApplier` implemented `TypedEventApplier` but was located in the
`processing.scaling` package instead of the canonical `state.appliers`
package where all other `TypedEventApplier` implementations resides.
This also caused it to be outside the golden file protection scope,
leaving its state mutations untracked and unprotected against unintended changes.

Move `ScalingUpApplier` to `state.appliers` package and update
`Scale_SCALING_UP_v1.golden` to reflect its actual state mutations now
that the golden file mechanism covers it.
